backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0          # async SQLite driver для SQLAlchemy

# ─── Configuration ────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ─── AI ───────────────────────────────────────────────────────────────────────
google-generativeai==0.8.3  # Gemini AI Studio API

# ─── Google Docs API ──────────────────────────────────────────────────────────
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ─── Web Scraping ─────────────────────────────────────────────────────────────
beautifulsoup4==4.12.3      # Для парсинга статичных JD-страниц
requests==2.32.3
lxml==5.3.0

# ─── Utilities ────────────────────────────────────────────────────────────────
httpx==0.28.1               # Async HTTP клиент
python-multipart==0.0.19    # Form data support в FastAPI
